src/main/java/ConferenceRoom.java
import java.util.ArrayList;

public class ConferenceRoom {
    private int capacity;
    private ArrayList<Guest> guests;
    private String name;

    public ConferenceRoom(int capacity, String name){
        this.capacity = capacity;
        this.name = name;
        this.guests = new ArrayList<Guest>();
    }

    public String getName() {
        return this.name;
    }

    public int getCapacity() {
        return this.capacity;
    }

    public int guestsCount() {
        return this.guests.size();
    }

    public void addGuest(Guest guest) {
        guests.add(guest);
    }
}
